reorganize so karma actually runs
