--- integration_tests/benchmarks/noui/browserstack.json
@@ -1,11 +1,0 @@
-{
-    "test_path": "index.html",
-    "browsers": [
-        {
-            "browser": "chrome",
-            "browser_version": "latest",
-            "os": "OS X",
-            "os_version": "Mountain Lion"
-        }
-    ]
-}
