--- integration_tests/benchmarks/noui/index.html
@@ -1,5 +1,0 @@
-<div id="test"></div>
-<script>
-  console.log('hello world');
-  document.getElementById('test').innerHTML = 'JavaScript ran!';
-</script>
