+++ integration_tests/benchmarks/package.json
@@ -3,9 +3,6 @@
   "version": "0.1.2",
   "description": "Benchmark tools for TensorFlow.js",
   "private": true,
-  "dependencies": {
-    "chart.js": "^2.7.2"
-  },
   "repository": {
     "type": "git",
     "url": "https://github.com/tensorflow/tfjs-core"
@@ -13,15 +10,11 @@
   "devDependencies": {
     "@tensorflow/tfjs-core": "0.12.4",
     "@types/jasmine": "~2.5.53",
-    "@types/polymer": "~1.1.31",
-    "browserify": "^16.2.2",
     "browserstack-runner": "~0.7.0",
     "clang-format": "~1.2.2",
-    "d3-array": "~1.2.1",
-    "d3-format": "~1.2.1",
-    "d3-scale": "~1.0.7",
     "jasmine": "~3.1.0",
     "jasmine-core": "~3.1.0",
+    "firebase": "~5.3.0",
     "karma": "~2.0.2",
     "karma-browserstack-launcher": "~1.3.0",
     "karma-chrome-launcher": "~2.2.0",
