+++ integration_tests/benchmarks/pool_benchmarks.ts
@@ -17,8 +17,8 @@
 
 import * as tf from '@tensorflow/tfjs-core';
 
-import {BenchmarkTest} from './benchmark';
-import * as benchmark_util from './benchmark_util';
+import * as benchmark_util from './util';
+import {BenchmarkTest} from './util';
 
 const CPU_OP_RUNS = 1;
