+++ integration_tests/benchmarks/ui/benchmark_run_groups.ts
@@ -15,13 +15,45 @@
  * =============================================================================
  */
 
-import {BatchNormalization3DCPUBenchmark, BatchNormalization3DGPUBenchmark} from '../common/batchnormalization3d_benchmark';
-import {BenchmarkRun, BenchmarkRunGroup} from '../common/benchmark';
-import {ConvGPUBenchmark, ConvParams, DepthwiseConvParams, RegularConvParams} from '../common/conv_benchmarks';
-import {MatmulCPUBenchmark, MatmulGPUBenchmark} from '../common/matmul_benchmarks';
-import {PoolBenchmarkParams, PoolCPUBenchmark, PoolGPUBenchmark} from '../common/pool_benchmarks';
-import {ReductionOpsCPUBenchmark, ReductionOpsGPUBenchmark} from '../common/reduction_ops_benchmark';
-import {UnaryOpsCPUBenchmark, UnaryOpsGPUBenchmark} from '../common/unary_ops_benchmark';
+import {BatchNormalization3DCPUBenchmark, BatchNormalization3DGPUBenchmark} from '../batchnormalization3d_benchmark';
+import {ConvGPUBenchmark, ConvParams, DepthwiseConvParams, RegularConvParams} from '../conv_benchmarks';
+import {MatmulCPUBenchmark, MatmulGPUBenchmark} from '../matmul_benchmarks';
+import {PoolBenchmarkParams, PoolCPUBenchmark, PoolGPUBenchmark} from '../pool_benchmarks';
+import {ReductionOpsCPUBenchmark, ReductionOpsGPUBenchmark} from '../reduction_ops_benchmark';
+import {UnaryOpsCPUBenchmark, UnaryOpsGPUBenchmark} from '../unary_ops_benchmark';
+import {BenchmarkTest} from '../util';
+
+export interface BenchmarkRunGroup {
+  name: string;
+  // Min and max steps to run the benchmark test over.
+  min: number;
+  max: number;
+  // The size of the step to take between benchmark runs.
+  stepSize: number;
+  // A transformation of step to the size passed to the benchmark test.
+  stepToSizeTransformation?: (step: number) => number;
+  // Option parameters which is given to the benchmark test. (e.g. ops types)
+  options?: string[];
+  selectedOption?: string;
+  benchmarkRuns: BenchmarkRun[];
+  params: {[option: string]: {}};
+}
+
+export class BenchmarkRun {
+  name: string;
+  benchmarkTest: BenchmarkTest;
+
+  chartData: ChartData[];
+  constructor(name: string, benchmarkTest: BenchmarkTest) {
+    this.name = name;
+    this.benchmarkTest = benchmarkTest;
+    this.chartData = [];
+  }
+
+  clearChartData() {
+    this.chartData = [];
+  }
+}
 
 export function getRunGroups(): BenchmarkRunGroup[] {
   const groups: BenchmarkRunGroup[] = [];
