+++ integration_tests/benchmarks/ui/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../tsconfig",
   "include": [
-    "./*.ts"
+    "*.ts"
   ]
 }
