+++ integration_tests/benchmarks/util.ts
@@ -17,6 +17,13 @@
 
 import * as tf from '@tensorflow/tfjs-core';
 
+// Maximum number of time before CPU tests don't execute during the next round.
+export const LAST_RUN_CPU_CUTOFF_MS = 5000;
+
+export interface BenchmarkTest {
+  run(size: number, opType?: string, params?: {}): Promise<number>;
+}
+
 export async function warmupAndBenchmarkGPU(benchmark: () => tf.Tensor):
     Promise<number> {
   // Warmup.
